Move change handler classes under common folder
