+++ packages/database/src/analytics/index.js
@@ -3,5 +3,4 @@
  * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
  */
 
-export { AnalyticsRefresher } from './AnalyticsRefresher';
 export { MaterializedViewLogDatabaseModel } from './MaterializedViewLogDatabaseModel';
