+++ packages/database/src/changeHandlers/EntityHierarchyCacher.js
