+++ packages/database/src/changeHandlers/index.js
@@ -3,5 +3,6 @@
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
  */
 
+export { AnalyticsRefresher } from './AnalyticsRefresher';
 export { EntityHierarchyCacher } from './EntityHierarchyCacher';
 export { SurveyResponseOutdater } from './SurveyResponseOutdater';
